Add files via upload
All the initial files of the design were uploaded except script.js and upsc.html. Both of these files are empty as no scripting has yet commenced and the UPSC page is a prospective one and is blank as of now. Further changes to the code will only be done on GitHub with first checks and original coding on replit.